refactor code.
