+++ src/com/zestedesavoir/zestwriter/utils/readability/Utilities.java
@@ -6,6 +6,7 @@
  */
 package com.zestedesavoir.zestwriter.utils.readability;
 
+import com.zestedesavoir.zestwriter.utils.ZestWriterErrorHandler;
 import org.w3c.dom.Document;
 import org.xml.sax.InputSource;
 import org.xml.sax.SAXException;
@@ -220,30 +221,7 @@ public class Utilities {
             factory.setIgnoringComments(true);
             factory.setIgnoringElementContentWhitespace(true);
             DocumentBuilder builder = factory.newDocumentBuilder();
-            builder.setErrorHandler(new org.xml.sax.ErrorHandler() {
-
-                // ignore fatal errors (an exception is guaranteed)
-                public void fatalError(SAXParseException exception) {
-
-                    System.out.println("** Error" + ", line " + exception.getLineNumber() + ", uri " + exception.getSystemId());
-                    System.out.println("   " + exception.getMessage());
-                }
-
-                // treat validation errors as fatal
-                public void error(SAXParseException e) throws SAXParseException {
-
-                    System.out.println("** Error" + ", line " + e.getLineNumber() + ", uri " + e.getSystemId());
-                    System.out.println("   " + e.getMessage());
-                    throw e;
-                }
-
-                // dump warnings too
-                public void warning(SAXParseException err) {
-
-                    System.out.println("** Warning" + ", line " + err.getLineNumber() + ", uri " + err.getSystemId());
-                    System.out.println("   " + err.getMessage());
-                }
-            });
+            builder.setErrorHandler(new ZestWriterErrorHandler());
             InputSource inputSource = new InputSource(new StringReader(file));
             MIQuery = builder.parse(inputSource);
         } catch (SAXException sxe) {
@@ -282,33 +260,7 @@ public class Utilities {
             factory.setIgnoringComments(true);
             factory.setIgnoringElementContentWhitespace(true);
             DocumentBuilder builder = factory.newDocumentBuilder();
-            builder.setErrorHandler(new org.xml.sax.ErrorHandler() {
-
-                // ignore fatal errors (an exception is guaranteed)
-                @Override
-                public void fatalError(SAXParseException err) {
-                    writeErrorToConsole(err);
-                }
-
-                // treat validation errors as fatal
-                @Override
-                public void error(SAXParseException err) throws SAXParseException {
-                    writeErrorToConsole(err);
-                    throw err;
-                }
-
-                // dump warnings too
-                @Override
-                public void warning(SAXParseException err) {
-                    writeErrorToConsole(err);
-                }
-
-                private void writeErrorToConsole(SAXParseException  err) {
-                    System.out.println("** Warning" + ", line " + err.getLineNumber() + ", uri " + err.getSystemId());
-                    System.out.println("   " + err.getMessage());
-                }
-            });
-
+            builder.setErrorHandler(new ZestWriterErrorHandler());
             MIQuery = builder.newDocument();
         } catch (ParserConfigurationException | FactoryConfigurationError pce) {
             // Parser with specified options can't be built
